optimze code structure
