+++ gym_uav_c_functions/CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.10)
+cmake_minimum_required(VERSION 2.8)
 project(c_functions)
 
 set(CMAKE_CXX_STANDARD 14)
