+++ uav_problem/__init__.py
@@ -1,0 +1,5 @@
+from uav_problem.problem_config import set_config as set_problem_config
+from uav_problem.problem_config import get_config as get_problem_config
+
+from uav_problem.problem import MyProblem as UavProblem
+
